moved component out, refined funding source auto complete
